backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
scikit-learn==1.7.1
pandas==2.3.2
joblib==1.5.1
sentence-transformers==5.1.0
faiss-cpu==1.12.0
numpy==2.3.2
python-dotenv==1.1.1
